interesse und schwierigkeitsgrad in die wokloadberechnung einbezogen

+++ app/model/Module.js
@@ -19,9 +19,9 @@ Ext.define('studiplaner.model.Module', {
             { name: 'name', type: 'string' },
             { name: 'ects', type: 'int' },
             { name: 'sws', type: 'int' },
-            { name: 'workload', type: 'int'},
-            { name: 'interest', type: 'int' },   
-			{ name: 'severity', type: 'int' }
+            { name: 'workload', type: 'float'},
+            { name: 'interest', type: 'float' },   
+			{ name: 'severity', type: 'float' }
         ],
         
         validations: [{
